requirements.txt
streamlit
pytube
openai-whisper
langchain-openai
python-dotenv